move common func query package to utils package
