--- command/query/export_test.go
@@ -1,7 +1,0 @@
-package query
-
-// export for test
-
-var GetQueryBytes = getQueryBytes
-var GetCost = getCost
-var ConvertByteToTByte = convertByteToTByte
