+++ utils/export_test.go
@@ -1,0 +1,5 @@
+package utils
+
+// export for test
+
+var ConvertByteToTByte = convertByteToTByte
