+++ utils/utils.go
@@ -3,9 +3,14 @@ package utils
 import (
 	"fmt"
 	"os/exec"
+	"regexp"
+	"strconv"
 )
 
 var EPSILON = 0.00000001
+
+// ToDo: read from config file
+const COSTPERTB = 5.0
 
 // CombineStrings is combine slice strings
 func CombineStrings(strs []string, sep string) string {
@@ -42,4 +47,28 @@ func FloatEquals(a, b float64) bool {
 		return true
 	}
 	return false
+}
+
+// GetQueryBytes is get query byte size
+func GetQueryBytes(str string) (string, error) {
+	regex := regexp.MustCompile(`running this query will process (\d+) bytes of data.`)
+	queryBytes := regex.FindStringSubmatch(string(str))
+	if len(queryBytes) != 2 {
+		return "", fmt.Errorf("unexpected result: bytes count is %d, must be %d", len(queryBytes), 1)
+	}
+	return queryBytes[1], nil
+}
+
+// GetCost is get query cost
+func GetCost(size string) (float64, error) {
+	sizeInt, err := strconv.Atoi(size)
+	if err != nil {
+		return 0.0, err
+	}
+	tByte := convertByteToTByte(float64(sizeInt))
+	return COSTPERTB * tByte, nil
+}
+
+func convertByteToTByte(b float64) float64 {
+	return b / (1024 * 1024 * 1024 * 1024)
 }
