+++ utils/utils_test.go
@@ -5,6 +5,8 @@ import (
 	"reflect"
 	"testing"
 
+	"strconv"
+
 	"github.com/snackmgmg/drybq/utils"
 )
 
@@ -49,5 +51,39 @@ func TestFloatEquals(t *testing.T) {
 
 	if !utils.FloatEquals(target, 0.3) {
 		t.Fatalf("unexpected result: target is %f", target)
+	}
+}
+
+func TestConvertByteToTByte(t *testing.T) {
+	target := 1099511627776
+	estimate := 1.0
+	if r := utils.ConvertByteToTByte(float64(target)); !utils.FloatEquals(r, estimate) {
+		t.Fatalf("get: %f, want: %f", r, estimate)
+	}
+}
+
+func TestGetCost(t *testing.T) {
+	target := strconv.Itoa(1099511627776 * 2)
+	estimate := 10.0
+
+	cost, err := utils.GetCost(target)
+	if err != nil {
+		t.Fatalf("error raised: %v", err)
+	}
+	if r := cost; !utils.FloatEquals(r, estimate) {
+		t.Fatalf("get: %f, want: %f", r, estimate)
+	}
+}
+
+func TestGetQueryBytes(t *testing.T) {
+	targetStr := `running this query will process 12345678 bytes of data.`
+	estimate := "12345678"
+
+	qBytes, err := utils.GetQueryBytes(targetStr)
+	if err != nil {
+		t.Fatalf("error raised: %v", err)
+	}
+	if r := qBytes; r != estimate {
+		t.Fatalf("get: %s, want: %s", r, estimate)
 	}
 }
